FO-2090: Fjern redux-form
Endelig er vi kvitt redux-form :)
