+++ package.json
@@ -80,13 +80,10 @@
     "react-modal": "3.5.1",
     "react-motion": "0.4.8",
     "react-redux": "^5.0.4",
-    "react-redux-form-validation": "1.1.0",
     "react-router-dom": "^5.0.1",
-    "react-router-redux": "4.0.8",
     "react-scripts": "3.0.1",
     "rebound": "0.0.15",
     "redux": "3.6.0",
-    "redux-form": "6.7.0",
     "redux-freeze": "0.1.5",
     "redux-thunk": "2.2.0",
     "reselect": "3.0.1",
