+++ src/felles-komponenter/skjema/field-group/fieldgroups-validering.less
