+++ src/felles-komponenter/skjema/input/input.js
@@ -1,49 +1,27 @@
+/* eslint-disable no-unused-vars */
 import React from 'react';
 import PT from 'prop-types';
 import { Input as NavInput } from 'nav-frontend-skjema';
-import { CustomField } from 'react-redux-form-validation';
-import { FormattedMessage } from 'react-intl';
 
-// eslint-disable-next-line no-unused-vars
-function InnerInputComponent({ input, labelId, errorMessage, meta, ...rest }) {
-    const feil = errorMessage ? { feilmelding: errorMessage[0] } : undefined;
+// pristine and initialValue isn't used, but we don't want to pass it to input
+function Input({ touched, error, input, pristine, initialValue, ...rest }) {
+    const feil = error && touched ? { feilmelding: error } : undefined;
     const inputProps = { ...input, ...rest };
-    return (
-        <NavInput
-            label={<FormattedMessage id={labelId} />}
-            feil={feil}
-            {...inputProps}
-        />
-    );
-}
-
-InnerInputComponent.propTypes = {
-    labelId: PT.string.isRequired,
-    bredde: PT.string,
-    errorMessage: PT.arrayOf(PT.node),
-    meta: PT.object, // eslint-disable-line react/forbid-prop-types
-    input: PT.object, // eslint-disable-line react/forbid-prop-types
-};
-
-InnerInputComponent.defaultProps = {
-    bredde: undefined,
-    errorMessage: undefined,
-    meta: undefined,
-    input: undefined,
-};
-
-function Input({ feltNavn, ...rest }) {
-    return (
-        <CustomField
-            name={feltNavn}
-            errorClass="skjemaelement--harFeil"
-            customComponent={<InnerInputComponent {...rest} />}
-        />
-    );
+    return <NavInput {...inputProps} feil={feil} />;
 }
 
 Input.propTypes = {
-    feltNavn: PT.string.isRequired,
+    initialValue: PT.string,
+    pristine: PT.bool,
+    touched: PT.bool.isRequired,
+    error: PT.string,
+    input: PT.object.isRequired,
+};
+
+Input.defaultProps = {
+    initialValue: undefined,
+    pristine: undefined,
+    error: undefined,
 };
 
 export default Input;
