--- src/moduler/aktivitet/aktivitet-forms/aktivitet-form-utils.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const formNavn = 'ny-aktivitet-form';
