--- src/moduler/aktivitet/aktivitet-forms/mote/mote-aktivitet-form.less
@@ -1,10 +1,0 @@
-.mote-aktivitet-form {
-    &__velg-mote-klokkeslett {
-        display: flex;
-        justify-content: space-between;
-
-        .input--xs {
-            width: 85px;
-        }
-    }
-}
