--- src/moduler/aktivitet/aktivitet-forms/samtalereferat/samtalereferat-form.less
@@ -1,3 +1,0 @@
-.samtalereferat-form__lagre-og-publiser {
-    margin-right: 1rem;
-}
