+++ src/moduler/aktivitet/aktivitet-forms/stilling/aktivitet-stilling-form.js
@@ -14,12 +14,12 @@ import {
     validateLenke,
     validateTittel,
 } from './validate';
-import Input from '../../../../felles-komponenter/skjema/input-v2/input';
+import Input from '../../../../felles-komponenter/skjema/input/input';
 import PeriodeValidering, {
     validerPeriodeFelt,
-} from '../../../../felles-komponenter/skjema/field-group/periode-valideringv2';
-import DatoField from '../../../../felles-komponenter/skjema/datovelger/datovelgerv2';
-import Textarea from '../../../../felles-komponenter/skjema/input-v2/textarea';
+} from '../../../../felles-komponenter/skjema/field-group/periode-validering';
+import DatoField from '../../../../felles-komponenter/skjema/datovelger/datovelger';
+import Textarea from '../../../../felles-komponenter/skjema/input/textarea';
 import { todayIsoString } from '../../../../utils';
 
 function StillingAktivitetForm(props) {
