+++ src/moduler/aktivitet/avslutt/begrunnelse-form.js
@@ -6,7 +6,7 @@ import useFormstate from '@nutgaard/use-formstate';
 import ModalContainer from '../../../felles-komponenter/modal/modal-container';
 import ModalFooter from '../../../felles-komponenter/modal/modal-footer';
 import FormErrorSummary from '../../../felles-komponenter/skjema/form-error-summary/form-error-summary';
-import Textarea from '../../../felles-komponenter/skjema/input-v2/textarea';
+import Textarea from '../../../felles-komponenter/skjema/input/textarea';
 
 const begrunnelseValidator = val => {
     if (val.trim().length === 0) {
